Move some code in pkg/test/testutil into dedicated packages
