+++ pkg/storage/conformance/token_store.go
@@ -7,10 +7,10 @@ import (
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
+	"github.com/rancher/opni/pkg/test/testruntime"
 
 	corev1 "github.com/rancher/opni/pkg/apis/core/v1"
 	"github.com/rancher/opni/pkg/storage"
-	"github.com/rancher/opni/pkg/test/testutil"
 	"github.com/rancher/opni/pkg/util/future"
 )
 
@@ -202,7 +202,7 @@ func TokenStoreTestSuite[T storage.TokenStore](
 
 				wg := sync.WaitGroup{}
 				start := make(chan struct{})
-				count := testutil.IfCI(3).Else(5)
+				count := testruntime.IfCI(3).Else(5)
 				for i := 0; i < count; i++ {
 					wg.Add(1)
 					go func() {
