+++ pkg/test/testutil/testing.go
@@ -2,46 +2,12 @@ package testutil
 
 import (
 	"errors"
-	"io"
-	"os"
 	"os/exec"
-	"strings"
 
 	"github.com/onsi/ginkgo/v2"
 	"github.com/onsi/gomega/gexec"
-	"github.com/samber/lo"
+	"github.com/rancher/opni/pkg/test/testruntime"
 )
-
-var (
-	IsTesting       = strings.HasSuffix(os.Args[0], ".test")
-	IsGithubActions = os.Getenv("GITHUB_ACTIONS") == "true"
-	IsDrone         = os.Getenv("DRONE") == "true"
-	StdoutWriter    = lo.Ternary[io.Writer](IsTesting, ginkgo.GinkgoWriter, os.Stdout)
-	StderrWriter    = lo.Ternary[io.Writer](IsTesting, ginkgo.GinkgoWriter, os.Stderr)
-)
-
-type ifExpr[T any] struct {
-	cond  bool
-	value T
-}
-
-func IfCI[T any](t T) ifExpr[T] {
-	return ifExpr[T]{
-		cond:  IsGithubActions || IsDrone,
-		value: t,
-	}
-}
-
-func (e ifExpr[T]) Else(t T) T {
-	return lo.Ternary(e.cond, e.value, t)
-}
-
-func IfTesting[T any](t T) ifExpr[T] {
-	return ifExpr[T]{
-		cond:  IsTesting,
-		value: t,
-	}
-}
 
 type Session interface {
 	G() (*gexec.Session, bool)
@@ -75,7 +41,7 @@ func (s *sessionWrapper) Wait() error {
 }
 
 func StartCmd(cmd *exec.Cmd) (Session, error) {
-	if IsTesting {
+	if testruntime.IsTesting {
 		session, err := gexec.Start(cmd, ginkgo.GinkgoWriter, ginkgo.GinkgoWriter)
 		if err != nil {
 			return nil, err
@@ -85,8 +51,8 @@ func StartCmd(cmd *exec.Cmd) (Session, error) {
 			cmd: cmd,
 		}, nil
 	}
-	cmd.Stdout = StdoutWriter
-	cmd.Stderr = StderrWriter
+	cmd.Stdout = testruntime.StdoutWriter
+	cmd.Stderr = testruntime.StderrWriter
 	if err := cmd.Start(); err != nil {
 		return nil, err
 	}
